==> src/11_Filter_data_na.R <==
df_pt <- read.csv("./data/processed/ab-ebdusers-person-trips.csv")

df_pt1 <- subset(df_pt,observer_id!="obsr292010")
df_pt1 <- subset(df_pt1,observer_id!="obsr136164")
df_pt1 <- subset(df_pt1,observer_id!="obsr326298")
df_pt1 <- subset(df_pt1,observer_id!="obsr406443")
df_pt1 <- subset(df_pt1,observer_id!="obsr413973")
df_pt1 <- subset(df_pt1,observer_id!="obsr534409")
df_pt1 <- subset(df_pt1,observer_id!="obsr629915")
df_pt1 <- subset(df_pt1,observer_id!="obsr685074")
df_pt1 <- subset(df_pt1,observer_id!="obsr715445")
df_pt1 <- subset(df_pt1,observer_id!="obsr250419")
df_pt1 <- subset(df_pt1,observer_id!="obsr352462")
df_pt1 <- subset(df_pt1,observer_id!="obsr410940")
df_pt1 <- subset(df_pt1,observer_id!="obsr418531")
df_pt1 <- subset(df_pt1,observer_id!="obsr418531")
df_pt1 <- subset(df_pt1,observer_id!="obsr552960")
df_pt1 <- subset(df_pt1,observer_id!="obsr611790")
df_pt1 <- subset(df_pt1,observer_id!="obsr683057")
df_pt1 <- subset(df_pt1,observer_id!="obsr803562")
df_pt1 <- subset(df_pt1,observer_id!="obsr821611")
df_pt1 <- subset(df_pt1,observer_id!="obsr116402")
df_pt1 <- subset(df_pt1,observer_id!="obsr164698")
df_pt1 <- subset(df_pt1,observer_id!="obsr179637")
df_pt1 <- subset(df_pt1,observer_id!="obsr198491")
df_pt1 <- subset(df_pt1,observer_id!="obsr397087")
df_pt1 <- subset(df_pt1,observer_id!="obsr419341")
df_pt1 <- subset(df_pt1,observer_id!="obsr548938")
df_pt1 <- subset(df_pt1,observer_id!="obsr654154")
df_pt1 <- subset(df_pt1,observer_id!="obsr723823")
df_pt1 <- subset(df_pt1,observer_id!="obsr816570")
df_pt1 <- subset(df_pt1,observer_id!="obsr816789")
df_pt1 <- subset(df_pt1,observer_id!="obsr823005")
df_pt1 <- subset(df_pt1,observer_id!="obsr823006")
df_pt1 <- subset(df_pt1,observer_id!="obsr823009")
df_pt1 <- subset(df_pt1,observer_id!="obsr835295")
df_pt1 <- subset(df_pt1,observer_id!="obsr932389")
df_pt1 <- subset(df_pt1,observer_id!="obsr939443")
df_pt1 <- subset(df_pt1,observer_id!="obsr979401")
df_pt1 <- subset(df_pt1,observer_id!="obsr544399")
df_pt1 <- subset(df_pt1,observer_id!="obsr148917")
df_pt1 <- subset(df_pt1,observer_id!="obsr227600")
df_pt1 <- subset(df_pt1,observer_id!="obsr284774")
df_pt1 <- subset(df_pt1,observer_id!="obsr289749")
df_pt1 <- subset(df_pt1,observer_id!="obsr330165")
df_pt1 <- subset(df_pt1,observer_id!="obsr369291")
df_pt1 <- subset(df_pt1,observer_id!="obsr512351")
df_pt1 <- subset(df_pt1,observer_id!="obsr823328")
df_pt1 <- subset(df_pt1,observer_id!="obsr920689")
df_pt1 <- subset(df_pt1,observer_id!="obsr949062")
df_pt1 <- subset(df_pt1,observer_id!="obsr286443")
df_pt1 <- subset(df_pt1,observer_id!="obsr1008025")
df_pt1 <- subset(df_pt1,observer_id!="obsr1008025")
df_pt1 <- subset(df_pt1,observer_id!="obsr125235")
df_pt1 <- subset(df_pt1,observer_id!="obsr347306")
df_pt1 <- subset(df_pt1,observer_id!="obsr675500")
df_pt1 <- subset(df_pt1,observer_id!="obsr776440")
df_pt1 <- subset(df_pt1,observer_id!="obsr970593")
df_pt1 <- subset(df_pt1,observer_id!="obsr138507")
df_pt1 <- subset(df_pt1,observer_id!="obsr299407")
df_pt1 <- subset(df_pt1,observer_id!="obsr391787")
df_pt1 <- subset(df_pt1,observer_id!="obsr429471")
df_pt1 <- subset(df_pt1,observer_id!="obsr431761")
df_pt1 <- subset(df_pt1,observer_id!="obsr552540")
df_pt1 <- subset(df_pt1,observer_id!="obsr679285")
df_pt1 <- subset(df_pt1,observer_id!="obsr730437")
df_pt1 <- subset(df_pt1,observer_id!="obsr748653")
df_pt1 <- subset(df_pt1,observer_id!="obsr1004879")
df_pt1 <- subset(df_pt1,observer_id!="obsr1010203")
df_pt1 <- subset(df_pt1,observer_id!="obsr148096")
df_pt1 <- subset(df_pt1,observer_id!="obsr198271")
df_pt1 <- subset(df_pt1,observer_id!="obsr211272")
df_pt1 <- subset(df_pt1,observer_id!="obsr221328")
df_pt1 <- subset(df_pt1,observer_id!="obsr272594")
df_pt1 <- subset(df_pt1,observer_id!="obsr313355")
df_pt1 <- subset(df_pt1,observer_id!="obsr317125")
df_pt1 <- subset(df_pt1,observer_id!="obsr649185")
df_pt1 <- subset(df_pt1,observer_id!="obsr660301")
df_pt1 <- subset(df_pt1,observer_id!="obsr738339")
df_pt1 <- subset(df_pt1,observer_id!="obsr760441")
df_pt1 <- subset(df_pt1,observer_id!="obsr789626")
df_pt1 <- subset(df_pt1,observer_id!="obsr803493")
df_pt1 <- subset(df_pt1,observer_id!="obsr1008269")
df_pt1 <- subset(df_pt1,observer_id!="obsr198287")
df_pt1 <- subset(df_pt1,observer_id!="obsr284705")
df_pt1 <- subset(df_pt1,observer_id!="obsr293784")
df_pt1 <- subset(df_pt1,observer_id!="obsr316462")
df_pt1 <- subset(df_pt1,observer_id!="obsr338028")
df_pt1 <- subset(df_pt1,observer_id!="obsr344617")
df_pt1 <- subset(df_pt1,observer_id!="obsr387308")
df_pt1 <- subset(df_pt1,observer_id!="obsr407552")
df_pt1 <- subset(df_pt1,observer_id!="obsr424835")
df_pt1 <- subset(df_pt1,observer_id!="obsr503058")
df_pt1 <- subset(df_pt1,observer_id!="obsr503585")
df_pt1 <- subset(df_pt1,observer_id!="obsr519218")
df_pt1 <- subset(df_pt1,observer_id!="obsr520258")
df_pt1 <- subset(df_pt1,observer_id!="obsr532157")
df_pt1 <- subset(df_pt1,observer_id!="obsr546276")
df_pt1 <- subset(df_pt1,observer_id!="obsr554898")
df_pt1 <- subset(df_pt1,observer_id!="obsr556862")
df_pt1 <- subset(df_pt1,observer_id!="obsr563729")
df_pt1 <- subset(df_pt1,observer_id!="obsr617589")
df_pt1 <- subset(df_pt1,observer_id!="obsr620566")
df_pt1 <- subset(df_pt1,observer_id!="obsr634738")
df_pt1 <- subset(df_pt1,observer_id!="obsr642391")
df_pt1 <- subset(df_pt1,observer_id!="obsr652326")
df_pt1 <- subset(df_pt1,observer_id!="obsr673173")
df_pt1 <- subset(df_pt1,observer_id!="obsr676111")
df_pt1 <- subset(df_pt1,observer_id!="obsr681884")
df_pt1 <- subset(df_pt1,observer_id!="obsr683708")
df_pt1 <- subset(df_pt1,observer_id!="obsr694517")
df_pt1 <- subset(df_pt1,observer_id!="obsr710818")
df_pt1 <- subset(df_pt1,observer_id!="obsr719736")
df_pt1 <- subset(df_pt1,observer_id!="obsr739198")
df_pt1 <- subset(df_pt1,observer_id!="obsr803504")
df_pt1 <- subset(df_pt1,observer_id!="obsr844067")
df_pt1 <- subset(df_pt1,observer_id!="obsr483922")
df_pt1 <- subset(df_pt1,observer_id!="obsr488321")
df_pt1 <- subset(df_pt1,observer_id!="obsr490084")
df_pt1 <- subset(df_pt1,observer_id!="obsr552989")
df_pt1 <- subset(df_pt1,observer_id!="obsr620287")
#df_pt1 <- subset(df_pt1, observer_id!="obsr283574") # outlier
#df_pt1 <- subset(df_pt1, observer_id!="obsr24588") # outlier


write.csv(df_pt1,"./data/processed/ab-ebdusers-person-trips1.csv")
